♻️ Refactor rest of examples to their own files
