+++ test/unit/mods.example.cet.ts
@@ -1,0 +1,149 @@
+import path from "path";
+import { InstallerType } from "../../src/installers.types";
+import { InstallChoices } from "../../src/ui.dialogs";
+import {
+  ExampleSucceedingMod,
+  CET_PREFIXES,
+  CET_PREFIX,
+  CET_INIT,
+  ExamplePromptInstallableMod,
+  copiedToSamePath,
+  ExamplesForType,
+  mergeOrFailOnConflict,
+  ExampleFailingMod,
+} from "./utils.helper";
+
+const CetModSucceeds = new Map<string, ExampleSucceedingMod>(
+  Object.entries({
+    cetWithOnlyInitCanonical: {
+      expectedInstallerType: InstallerType.CET,
+      inFiles: [
+        ...CET_PREFIXES,
+        path.join(`${CET_PREFIX}/exmod/`),
+        path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+      ],
+      outInstructions: [
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+          destination: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+        },
+      ],
+    },
+    cetWithTypicalMultipleFilesCanonical: {
+      expectedInstallerType: InstallerType.CET,
+      inFiles: [
+        ...CET_PREFIXES,
+        path.join(`${CET_PREFIX}/exmod/`),
+        path.join(`${CET_PREFIX}/exmod/AdditionalSubFolder/`),
+        path.join(`${CET_PREFIX}/exmod/Modules/`),
+        path.join(`${CET_PREFIX}/exmod/configfile.json`),
+        path.join(`${CET_PREFIX}/exmod/db.sqlite3`),
+        path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+        path.join(`${CET_PREFIX}/exmod/README.md`),
+        path.join(`${CET_PREFIX}/exmod/AdditionalSubFolder/Whoaonemore/`),
+        path.join(`${CET_PREFIX}/exmod/AdditionalSubFolder/Whoaonemore/init.lua`),
+        path.join(`${CET_PREFIX}/exmod/AdditionalSubFolder/strangestuff.lua`),
+        path.join(`${CET_PREFIX}/exmod/Modules/UI.lua`),
+        path.join(`${CET_PREFIX}/exmod/Modules/MagicCheats.lua`),
+      ],
+      outInstructions: [
+        {
+          type: `copy`,
+          source: path.join(
+            `${CET_PREFIX}/exmod/AdditionalSubFolder/Whoaonemore/init.lua`,
+          ),
+          destination: path.join(
+            `${CET_PREFIX}/exmod/AdditionalSubFolder/Whoaonemore/init.lua`,
+          ),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/AdditionalSubFolder/strangestuff.lua`),
+          destination: path.join(
+            `${CET_PREFIX}/exmod/AdditionalSubFolder/strangestuff.lua`,
+          ),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/Modules/UI.lua`),
+          destination: path.join(`${CET_PREFIX}/exmod/Modules/UI.lua`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/Modules/MagicCheats.lua`),
+          destination: path.join(`${CET_PREFIX}/exmod/Modules/MagicCheats.lua`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/configfile.json`),
+          destination: path.join(`${CET_PREFIX}/exmod/configfile.json`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/db.sqlite3`),
+          destination: path.join(`${CET_PREFIX}/exmod/db.sqlite3`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+          destination: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/README.md`),
+          destination: path.join(`${CET_PREFIX}/exmod/README.md`),
+        },
+      ],
+    },
+    cetWithIniFilesCanonical: {
+      expectedInstallerType: InstallerType.CET,
+      inFiles: [
+        ...CET_PREFIXES,
+        path.join(`${CET_PREFIX}/exmod/`),
+        path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+        path.join(`${CET_PREFIX}/exmod/some.ini`),
+      ],
+      outInstructions: [
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+          destination: path.join(`${CET_PREFIX}/exmod/${CET_INIT}`),
+        },
+        {
+          type: `copy`,
+          source: path.join(`${CET_PREFIX}/exmod/some.ini`),
+          destination: path.join(`${CET_PREFIX}/exmod/some.ini`),
+        },
+      ],
+    },
+  }),
+);
+
+const CetModShouldPromptForInstall = new Map<string, ExamplePromptInstallableMod>(
+  Object.entries({
+    cetModWithIniShouldPromptToInstall: {
+      expectedInstallerType: InstallerType.Fallback,
+      inFiles: [
+        path.join(`exmod/`),
+        path.join(`exmod/${CET_INIT}`),
+        path.join(`exmod/some.ini`),
+      ],
+      proceedLabel: InstallChoices.Proceed,
+      proceedOutInstructions: [
+        copiedToSamePath(`exmod\\${CET_INIT}`),
+        copiedToSamePath(`exmod\\some.ini`),
+      ],
+      cancelLabel: InstallChoices.Cancel,
+      cancelErrorMessage: `Fallback Installer: user chose to cancel installation`,
+    },
+  }),
+);
+
+const examples: ExamplesForType = {
+  AllExpectedSuccesses: mergeOrFailOnConflict(CetModSucceeds),
+  AllExpectedDirectFailures: new Map<string, ExampleFailingMod>(),
+  AllExpectedPromptInstalls: mergeOrFailOnConflict(CetModShouldPromptForInstall),
+};
+
+export default examples;
